pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waydeeper"
version = "0.1.0"
description = "GPU-accelerated depth effect wallpaper for Wayland"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "numpy",
    "pillow",
    "onnxruntime",
    "PyGObject",
    "PyOpenGL",
    "PyOpenGL-accelerate",
]

[project.scripts]
waydeeper = "src.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
